Login and Signup
Added methods and basic pages for logging in and signing up. Needs some work.
Until we fully set up user authentication (as in preventing access to pages etc), some things have been left as-is to avoid testing issues (that is, to avoid having to create a user everytime we want to run the program):
-userID should eventually be set to false
-all methods that retrieve list of trips should eventually call the findAllByUserIDOrderByStartAsc method
We'll need to eventually remap some things as well. For now, a link to the login and sign up pages is available on the homepage via the "access account" button.
Right now, successfully logging in will just return to homepage.
Feel free to add to User class info and login/signup pages.
Also fixed some bugs introduced by integrating 'destinations' fields.